deletSingleBooking edited with validation

+++ src/controllers/booking.controller.js
@@ -1,5 +1,5 @@
 const bookingsCollection = require("../models/booking.model");
-const {ObjectId} = require("mongodb");
+const { ObjectId } = require("mongodb");
 
 exports.getAllBooking = async (req, res) => {
     const result = await bookingsCollection.find().toArray();
@@ -7,8 +7,8 @@ exports.getAllBooking = async (req, res) => {
 }
 
 exports.getUserSpecificBooking = async (req, res) => {
-    const {email} = req.params;
-    const result = await bookingsCollection.find({email}).toArray();
+    const { email } = req.params;
+    const result = await bookingsCollection.find({ email }).toArray();
     res.send(result);
 }
 
@@ -20,8 +20,29 @@ exports.createBooking = async (req, res) => {
 
 
 exports.deletSingleBooking = async (req, res) => {
-    const {id} = req.params;
-    const query = {_id : new ObjectId(id)};
-    const result = await bookingsCollection.deleteOne(query);
-    res.send(result);
+    try {
+        const { id } = req.params;
+        const { email: loggedInEmail } = req.body;
+        if (!loggedInEmail) {
+            return res.status(401).send({ message: "Unauthorized" });
+        }
+        // find the booking first
+        const booking = await bookingsCollection.findOne({ _id: new ObjectId(id) });
+
+        if (!booking) {
+            return res.status(404).send({ message: "Booking not found" });
+        }
+
+        // check if the logged in user is the owner of the booking
+        if (booking.email !== loggedInEmail) {
+            return res.status(403).send({ message: "Unauthorized: This is not your booking" });
+        }
+
+        const query = { _id: new ObjectId(id) };
+        const result = await bookingsCollection.deleteOne(query);
+        res.send(result);
+    } catch (error) {
+        console.log(error);
+        res.status(500).send({ message: "Internal server error" });
+    }
 }
